+++ openmbvcppinterface/openmbvcppinterface/flexiblebody.h
@@ -31,7 +31,7 @@ namespace OpenMBV {
   class FlexibleBody : public DynamicColoredBody {
     protected:
       int numvp{0};
-      H5::VectorSerie<double>* data;
+      H5::VectorSerie<Float>* data;
       FlexibleBody() = default;
       ~FlexibleBody() override = default;
       xercesc::DOMElement *writeXMLFile(xercesc::DOMNode *parent) override;
@@ -54,7 +54,7 @@ namespace OpenMBV {
       }
 
       int getRows() override { return data?data->getRows():0; }
-      std::vector<double> getRow(int i) override { return data?data->getRow(i):std::vector<double>(1+3*numvp); }
+      std::vector<Float> getRow(int i) override { return data?data->getRow(i):std::vector<Float>(1+3*numvp); }
 
       /** Initializes the time invariant part of the object using a XML node */
       void initializeUsingXML(xercesc::DOMElement *element) override;
